--- src/Database/Repositories/Revenue.php
@@ -1,10 +1,0 @@
-<?php
-namespace Give\Database\Repositories;
-
-/**
- * Class Revenue
- * @package Give\Database\Repositories
- *
- * @since 2.9.0
- */
-class Revenue extends TableAccessor{}
