+++ src/Revenue/Repositories/Revenue.php
@@ -1,0 +1,12 @@
+<?php
+namespace Give\Revenue\Repositories;
+
+use Give\Framework\TableAccessor;
+
+/**
+ * Class Revenue
+ * @package Give\Database\Repositories
+ *
+ * @since 2.9.0
+ */
+class Revenue extends TableAccessor {}
