--- src/ServiceProviders/DatabaseTables.php
@@ -1,54 +1,0 @@
-<?php
-namespace  Give\ServiceProviders;
-
-use Give\Database\Tables\Revenue;
-use Give\Database\Repositories\Revenue as RevenueRepository;
-
-/**
- * Class DatabaseTables
- * @package Give\ServiceProviders
- *
- * @since 2.9.0
- */
-class DatabaseTables implements ServiceProvider {
-	/**
-	 * Table list.
-	 * @since 2.9.0
-	 * @var string[]
-	 */
-	private $tables = [
-		Revenue::class,
-	];
-
-	/**
-	 * Table list.
-	 * @since 2.9.0
-	 * @var string[]
-	 */
-	private $repositories = [
-		Revenue::class => RevenueRepository::class,
-	];
-
-	/**
-	 * @inheritDoc
-	 */
-	public function register() {
-		foreach ( $this->tables as $table ) {
-			give()->singleton( $table );
-		}
-
-		foreach ( $this->repositories as $table => $repository ) {
-			give()->bind(
-				$repository,
-				function() use ( $repository, $table ) {
-					new $repository( give( $table ) );
-				}
-			);
-		}
-	}
-
-	/**
-	 * @inheritdoc
-	 */
-	public function boot() {}
-}
